# docs/api/openapi.yaml
openapi: 3.0.3
info:
  version: 1.0.0
  title: Agent Banking Platform API
  description: Aggregated API from all microservices
  contact:
    name: Agent Banking Platform Team
servers:
- description: Gateway
  url: http://localhost:8080
- description: Production
  url: https://api.agentbanking.example.com
- description: Sandbox
  url: https://sandbox.agentbanking.example.com
paths:
  /api/v1/transactions/quote:
    post:
      tags:
      - transaction-controller (rules-service)
      operationId: getTransactionQuote
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransactionQuoteRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionQuoteResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/rules/fees:
    post:
      tags:
      - rules-controller (rules-service)
      operationId: createFeeConfig
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeeConfigRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FeeConfigResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/backoffice/audit-logs:
    get:
      tags:
      - audit-log-controller (onboarding-service)
      operationId: getAuditLogs
      security:
        - bearerAuth: []
      parameters:
      - name: entityType
        in: query
        required: false
        schema:
          type: string
      - name: fromDate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: toDate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 50
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: object
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    TransactionQuoteRequest:
      required:
      - amount
      - serviceCode
      - fundingSource
      - agentId
      type: object
      properties:
        serviceCode:
          type: string
          enum: [CASH_WITHDRAWAL, CASH_DEPOSIT, BILL_PAYMENT, DUITNOW_TRANSFER, CASHLESS_PAYMENT, PIN_BASED_PURCHASE, PREPAID_TOPUP, EWALLET_WITHDRAWAL, EWALLET_TOPUP, ESSP_PURCHASE, PIN_PURCHASE, RETAIL_SALE, HYBRID_CASHBACK, BALANCE_INQUIRY, JOMPAY, ASTRO_RPN, TM_RPN, CELCOM_TOPUP, M1_TOPUP]
          description: Service code for the transaction. Legacy codes (JOMPAY, ASTRO_RPN, TM_RPN, CELCOM_TOPUP, M1_TOPUP) are mapped to canonical types.
        amount:
          type: string
        agentId:
          type: string
        fundingSource:
          type: string
          enum: [CARD_EMV, CASH, DUITNOW_MOBILE, DUITNOW_MYKAD, DUITNOW_BRN, MYKAD_BIOMETRIC, DUITNOW_QR]
        billerRouting:
          type: string
          enum: [ON_US, OFF_US]
    TransactionQuoteResponse:
      required:
      - amount
      - fee
      - commission
      - total
      - quoteId
      type: object
      properties:
        quoteId:
          type: string
        amount:
          type: string
        fee:
          type: string
        total:
          type: string
        commission:
          type: string
...
    FeeConfigRequest:
      required:
      - agentTier
      - transactionType
      - feeType
      - currency
      - effectiveFrom
      type: object
      properties:
        agentTier:
          type: string
          enum: [MICRO, STANDARD, PREMIER]
        transactionType:
          type: string
          enum: [CASH_WITHDRAWAL, CASH_DEPOSIT, BALANCE_INQUIRY, DUITNOW_TRANSFER, BILL_PAYMENT, JOMPAY, ASTRO_RPN, TM_RPN, EPF_PAYMENT, PREPAID_TOPUP, CELCOM_TOPUP, M1_TOPUP, SARAWAK_PAY_WITHDRAWAL, SARAWAK_PAY_TOPUP, ESSP_PURCHASE, PIN_PURCHASE, CASHLESS_PAYMENT, PIN_BASED_PURCHASE, EWALLET_WITHDRAWAL, EWALLET_TOPUP, RETAIL_SALE, HYBRID_CASHBACK]
          description: Transaction type. JOMPAY, ASTRO_RPN, TM_RPN map to BILL_PAYMENT. CELCOM_TOPUP, M1_TOPUP map to PREPAID_TOPUP.
        feeType:
          type: string
          enum: [FIXED, PERCENTAGE]
        feeAmount:
          type: string
        percentage:
          type: string
        minFee:
          type: string
        maxFee:
          type: string
        currency:
          type: string
          enum: [MYR]
        effectiveFrom:
          type: string
          format: date-time
        effectiveTo:
          type: string
          format: date-time
    
    # Discrepancy Maker Action Request
...
    FeeConfigResponse:
      type: object
      properties:
        feeConfigId:
          type: string
          format: uuid
        agentTier:
          type: string
        transactionType:
          type: string
        feeType:
          type: string
        feeAmount:
          type: string
        percentage:
          type: string
        status:
          type: string
          enum: [CREATED, ACTIVE, INACTIVE]
        effectiveFrom:
          type: string
          format: date-time
        effectiveTo:
          type: string
          format: date-time
    
    # KYC Verify Response
...
    ErrorResponse:
      required:
      - status
      - error
      type: object
      properties:
        status:
          type: string
          enum: [FAILED]
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            action_code:
              type: string
              enum: [DECLINE, RETRY, REVIEW]
            trace_id:
              type: string
            timestamp:
              type: string
              format: date-time

    # ============================================================
    # TRANSACTION ORCHESTRATOR SCHEMAS (NEW - 2026-04-05)
    # ============================================================

    # Transaction Type Enum
...